modularize desktop api and routing
