+++ apps/desktop/src/shell/useDesktopShellRouting.ts
@@ -9,40 +9,32 @@ import {
 } from 'react';
 import { useLocation, useNavigate } from 'react-router-dom';
 
-import type {
-  DesktopApi,
-} from '@/lib/api';
+import type { DesktopApi } from '@/lib/api';
 
 import type {
   PrimarySection,
   ProfileConnectionsView,
-  SettingsSection,
   TimelineWorkspaceView,
 } from '@/components/shell/types';
 import {
-  PRIMARY_SECTION_PATHS,
-  type DesktopShellRouteOverrides,
   type OpenAuthorOptions,
   type OpenThreadOptions,
-  isProfileConnectionsView,
-  isSettingsSection,
   parsePrimarySectionPath,
   resolveHashBackedRouteLocation,
 } from '@/shell/routes';
+import { THREAD_TIMELINE_LIMIT } from '@/shell/pagination';
+import { useRouteSynchronization } from '@/shell/routing/useRouteSynchronization';
+import { useSyncRoute } from '@/shell/routing/useSyncRoute';
+import {
+  authorViewFromDirectMessageConversation,
+  mergeKnownAuthors,
+  messageFromError,
+} from '@/shell/selectors';
 import {
   useDesktopShellFieldSetter,
   useDesktopShellStore,
   useDesktopShellStoreApi,
 } from '@/shell/store';
-import { THREAD_TIMELINE_LIMIT } from '@/shell/pagination';
-import {
-  authorViewFromDirectMessageConversation,
-  isHex64,
-  mergeKnownAuthors,
-  messageFromError,
-  privateComposeTarget,
-  privateTimelineScope,
-} from '@/shell/selectors';
 
 type UseDesktopShellRoutingArgs = {
   api: DesktopApi;
@@ -71,29 +63,14 @@ export function useDesktopShellRouting({
   const state = useDesktopShellStore();
   const {
     activeTopic,
-    trackedTopics,
-    joinedChannelsByTopic,
-    selectedChannelIdByTopic,
-    liveSessionsByTopic,
-    gameRoomsByTopic,
-    livePanelStateByTopic,
-    gamePanelStateByTopic,
     selectedThread,
-    focusedObjectId,
-    thread,
     selectedAuthorPubkey,
-    selectedAuthor,
-    directMessagePaneOpen,
     selectedDirectMessagePeerPubkey,
-    selectedLiveSessionId,
-    selectedGameRoomId,
     lastNonNotificationsRoute,
     shellChromeState,
   } = state;
 
   const setActiveTopic = useDesktopShellFieldSetter('activeTopic');
-  const setComposeChannelByTopic = useDesktopShellFieldSetter('composeChannelByTopic');
-  const setSelectedChannelIdByTopic = useDesktopShellFieldSetter('selectedChannelIdByTopic');
   const setSelectedThread = useDesktopShellFieldSetter('selectedThread');
   const setFocusedObjectId = useDesktopShellFieldSetter('focusedObjectId');
   const setThread = useDesktopShellFieldSetter('thread');
@@ -113,7 +90,6 @@ export function useDesktopShellRouting({
   const setDirectMessageStatusByPeer = useDesktopShellFieldSetter('directMessageStatusByPeer');
   const setDirectMessageError = useDesktopShellFieldSetter('directMessageError');
   const setKnownAuthorsByPubkey = useDesktopShellFieldSetter('knownAuthorsByPubkey');
-  const setTimelineScopeByTopic = useDesktopShellFieldSetter('timelineScopeByTopic');
   const setSelectedLiveSessionId = useDesktopShellFieldSetter('selectedLiveSessionId');
   const setSelectedGameRoomId = useDesktopShellFieldSetter('selectedGameRoomId');
   const setError = useDesktopShellFieldSetter('error');
@@ -154,126 +130,12 @@ export function useDesktopShellRouting({
     };
   }, []);
 
-  const syncRoute = useCallback(
-    (mode: 'push' | 'replace' = 'replace', overrides?: DesktopShellRouteOverrides) => {
-      const currentState = storeApi.getState();
-      const hasOverride = <K extends keyof DesktopShellRouteOverrides>(key: K) =>
-        overrides ? Object.prototype.hasOwnProperty.call(overrides, key) : false;
-      const search = new URLSearchParams();
-      const nextTopic = overrides?.activeTopic ?? currentState.activeTopic;
-      const nextPrimarySection =
-        overrides?.primarySection ?? currentState.shellChromeState.activePrimarySection;
-      const nextTimelineView =
-        overrides?.timelineView ?? currentState.shellChromeState.timelineView;
-      const nextProfileMode =
-        overrides?.profileMode ?? currentState.shellChromeState.profileMode;
-      const nextProfileConnectionsView =
-        overrides?.profileConnectionsView ?? currentState.shellChromeState.profileConnectionsView;
-      const nextSelectedThread = hasOverride('selectedThread')
-        ? overrides?.selectedThread ?? null
-        : currentState.selectedThread;
-      const nextFocusedObjectId = hasOverride('focusedObjectId')
-        ? overrides?.focusedObjectId ?? null
-        : currentState.focusedObjectId;
-      const nextSelectedAuthorPubkey = hasOverride('selectedAuthorPubkey')
-        ? overrides?.selectedAuthorPubkey ?? null
-        : currentState.selectedAuthorPubkey;
-      const nextSelectedDirectMessagePeerPubkey = hasOverride('selectedDirectMessagePeerPubkey')
-        ? overrides?.selectedDirectMessagePeerPubkey ?? null
-        : currentState.selectedDirectMessagePeerPubkey;
-      const nextSelectedLiveSessionId = hasOverride('selectedLiveSessionId')
-        ? overrides?.selectedLiveSessionId ?? null
-        : currentState.selectedLiveSessionId;
-      const nextSelectedGameRoomId = hasOverride('selectedGameRoomId')
-        ? overrides?.selectedGameRoomId ?? null
-        : currentState.selectedGameRoomId;
-      const nextSettingsOpen = hasOverride('settingsOpen')
-        ? overrides?.settingsOpen ?? false
-        : currentState.shellChromeState.settingsOpen;
-      const nextSettingsSection =
-        overrides?.settingsSection ?? currentState.shellChromeState.activeSettingsSection;
-      let nextSelectedChannelId = currentState.selectedChannelIdByTopic[nextTopic] ?? null;
-
-      if (hasOverride('composeTarget')) {
-        nextSelectedChannelId =
-          overrides?.composeTarget?.kind === 'private_channel'
-            ? overrides.composeTarget.channel_id
-            : null;
-      } else if (hasOverride('timelineScope')) {
-        nextSelectedChannelId =
-          overrides?.timelineScope?.kind === 'channel'
-            ? overrides.timelineScope.channel_id
-            : null;
-      }
-
-      search.set('topic', nextTopic);
-      if (
-        nextPrimarySection !== 'messages' &&
-        nextPrimarySection !== 'notifications' &&
-        nextSelectedChannelId &&
-        !(nextPrimarySection === 'timeline' && nextTimelineView === 'bookmarks')
-      ) {
-        search.set('channel', nextSelectedChannelId);
-      }
-      if (nextPrimarySection === 'timeline' && nextTimelineView === 'bookmarks') {
-        search.set('timelineView', 'bookmarks');
-      }
-      if (nextPrimarySection === 'messages') {
-        if (nextSelectedDirectMessagePeerPubkey) {
-          search.set('peerPubkey', nextSelectedDirectMessagePeerPubkey);
-        }
-        if (nextSelectedAuthorPubkey) {
-          search.set('authorPubkey', nextSelectedAuthorPubkey);
-        }
-      } else if (nextPrimarySection !== 'notifications' && nextSelectedThread) {
-        search.set('context', 'thread');
-        search.set('threadId', nextSelectedThread);
-        if (nextFocusedObjectId) {
-          search.set('focusObjectId', nextFocusedObjectId);
-        }
-        if (nextSelectedAuthorPubkey) {
-          search.set('authorPubkey', nextSelectedAuthorPubkey);
-        }
-      } else if (nextPrimarySection !== 'notifications' && nextSelectedAuthorPubkey) {
-        search.set('context', 'author');
-        search.set('authorPubkey', nextSelectedAuthorPubkey);
-      }
-      if (nextPrimarySection === 'profile' && nextProfileMode === 'edit') {
-        search.set('profileMode', 'edit');
-      }
-      if (nextPrimarySection === 'profile' && nextProfileMode === 'connections') {
-        search.set('profileMode', 'connections');
-        search.set('connectionsView', nextProfileConnectionsView);
-      }
-      if (nextSettingsOpen) {
-        search.set('settings', nextSettingsSection);
-      }
-      if (nextPrimarySection === 'live' && nextSelectedLiveSessionId) {
-        search.set('sessionId', nextSelectedLiveSessionId);
-      }
-      if (nextPrimarySection === 'game' && nextSelectedGameRoomId) {
-        search.set('roomId', nextSelectedGameRoomId);
-      }
-
-      const nextPath = PRIMARY_SECTION_PATHS[nextPrimarySection];
-      const nextSearch = search.toString();
-      const nextUrl = nextSearch ? `${nextPath}?${nextSearch}` : nextPath;
-      const currentUrl = `${resolvedRouteLocation.pathname}${resolvedRouteLocation.search}`;
-      if (currentUrl !== nextUrl) {
-        pendingRouteUrlRef.current = nextUrl;
-        navigate(nextUrl, { replace: mode === 'replace' });
-        return;
-      }
-      pendingRouteUrlRef.current = null;
-    },
-    [
-      navigate,
-      pendingRouteUrlRef,
-      resolvedRouteLocation.pathname,
-      resolvedRouteLocation.search,
-      storeApi,
-    ]
-  );
+  const syncRoute = useSyncRoute({
+    navigate,
+    pendingRouteUrlRef,
+    resolvedRouteLocation,
+    storeApi,
+  });
 
   const setNavOpen = useCallback(
     (open: boolean, restoreToTrigger = false) => {
@@ -955,771 +817,21 @@ export function useDesktopShellRouting({
     setShellChromeState,
   ]);
 
-  useEffect(() => {
-    const currentUrl = `${resolvedRouteLocation.pathname}${resolvedRouteLocation.search}`;
-    const routeChanged = lastObservedRouteUrlRef.current !== currentUrl;
-    if (pendingRouteUrlRef.current && pendingRouteUrlRef.current !== currentUrl) {
-      if (!routeChanged) {
-        return;
-      }
-      pendingRouteUrlRef.current = null;
-    }
-    pendingRouteUrlRef.current = null;
-    lastObservedRouteUrlRef.current = currentUrl;
-    if (routeSection !== 'notifications') {
-      setLastNonNotificationsRoute(currentUrl);
-    }
-
-    if (!parsePrimarySectionPath(resolvedRouteLocation.pathname)) {
-      navigate(`${PRIMARY_SECTION_PATHS.timeline}${resolvedRouteLocation.search}`, {
-        replace: true,
-      });
-      return;
-    }
-
-    const params = new URLSearchParams(resolvedRouteLocation.search);
-    const requestedTopic = params.get('topic')?.trim() ?? null;
-    const requestedChannelParam = params.get('channel')?.trim() ?? null;
-    const requestedTimelineView = params.get('timelineView');
-    const requestedTimelineScopeValue = params.get('timelineScope');
-    const requestedComposeTargetValue = params.get('composeTarget');
-    const requestedSettingsSection = params.get('settings');
-    const requestedContext = params.get('context');
-    const requestedProfileMode = params.get('profileMode');
-    const requestedConnectionsView = params.get('connectionsView');
-    const requestedThreadId = params.get('threadId');
-    const requestedFocusObjectId = params.get('focusObjectId')?.trim() ?? null;
-    const requestedAuthorPubkey = params.get('authorPubkey');
-    const requestedPeerPubkey = params.get('peerPubkey');
-    const requestedSessionId = params.get('sessionId')?.trim() ?? null;
-    const requestedRoomId = params.get('roomId')?.trim() ?? null;
-
-    let nextTopic = activeTopic;
-    let shouldReload = false;
-    let shouldNormalize = false;
-    let normalizedSelectedThread: string | null = selectedThread;
-    let normalizedFocusedObjectId: string | null = focusedObjectId;
-    let normalizedSelectedAuthorPubkey: string | null = selectedAuthorPubkey;
-    let normalizedSelectedDirectMessagePeerPubkey: string | null =
-      selectedDirectMessagePeerPubkey;
-    let normalizedSelectedLiveSessionId: string | null = selectedLiveSessionId;
-    let normalizedSelectedGameRoomId: string | null = selectedGameRoomId;
-
-    if (requestedTopic) {
-      if (trackedTopics.includes(requestedTopic)) {
-        if (requestedTopic !== activeTopic) {
-          nextTopic = requestedTopic;
-          setActiveTopic(requestedTopic);
-          shouldReload = true;
-        }
-      } else {
-        shouldNormalize = true;
-      }
-    } else {
-      shouldNormalize = true;
-    }
-
-    const nextTimelineView =
-      routeSection === 'timeline' && requestedTimelineView === 'bookmarks' ? 'bookmarks' : 'feed';
-    const joinedChannelsForTopic = joinedChannelsByTopic[nextTopic] ?? [];
-    const liveSessionsForTopic = liveSessionsByTopic[nextTopic] ?? [];
-    const gameRoomsForTopic = gameRoomsByTopic[nextTopic] ?? [];
-    const livePanelState = livePanelStateByTopic[nextTopic];
-    const gamePanelState = gamePanelStateByTopic[nextTopic];
-    const currentSelectedChannelIdForTopic = selectedChannelIdByTopic[nextTopic] ?? null;
-    const allowChannelRouteParam =
-      routeSection !== 'messages' && routeSection !== 'notifications';
-    let nextSelectedChannelId = currentSelectedChannelIdForTopic;
-    if (allowChannelRouteParam && nextTimelineView !== 'bookmarks') {
-      nextSelectedChannelId = requestedChannelParam;
-      if (!nextSelectedChannelId) {
-        const legacyRequestedChannel = [requestedComposeTargetValue, requestedTimelineScopeValue]
-          .filter((value): value is string => Boolean(value))
-          .map((value) => {
-            if (value.startsWith('channel:')) {
-              return value.slice('channel:'.length);
-            }
-            return null;
-          })
-          .find((value): value is string => value !== null);
-        if (legacyRequestedChannel) {
-          nextSelectedChannelId = legacyRequestedChannel;
-        }
-      }
-    } else if (requestedChannelParam) {
-      shouldNormalize = true;
-    }
-    if (requestedTimelineScopeValue || requestedComposeTargetValue) {
-      shouldNormalize = true;
-    }
-    if (
-      nextTimelineView !== 'bookmarks' &&
-      nextSelectedChannelId &&
-      !joinedChannelsForTopic.some((channel) => channel.channel_id === nextSelectedChannelId)
-    ) {
-      shouldNormalize = true;
-      nextSelectedChannelId = null;
-    }
-
-    if (currentSelectedChannelIdForTopic !== nextSelectedChannelId) {
-      setSelectedChannelIdByTopic((current) => ({
-        ...current,
-        [nextTopic]: nextSelectedChannelId,
-      }));
-      setTimelineScopeByTopic((current) => ({
-        ...current,
-        [nextTopic]: privateTimelineScope(nextSelectedChannelId),
-      }));
-      setComposeChannelByTopic((current) => ({
-        ...current,
-        [nextTopic]: privateComposeTarget(nextSelectedChannelId),
-      }));
-      shouldReload = true;
-    }
-
-    if (requestedContext === 'dm' && routeSection !== 'messages') {
-      scheduleAnimationFrame(() => {
-        syncRoute('replace', {
-          activeTopic: nextTopic,
-          primarySection: 'messages',
-          selectedAuthorPubkey: null,
-          selectedDirectMessagePeerPubkey:
-            requestedPeerPubkey && isHex64(requestedPeerPubkey) ? requestedPeerPubkey : null,
-          selectedThread: null,
-        });
-      });
-      return;
-    }
-
-    const nextSettingsOpen = isSettingsSection(requestedSettingsSection);
-    const nextSettingsSection = isSettingsSection(requestedSettingsSection)
-      ? requestedSettingsSection
-      : shellChromeState.activeSettingsSection;
-    const nextProfileMode =
-      routeSection === 'profile'
-        ? requestedProfileMode === 'edit'
-          ? 'edit'
-          : requestedProfileMode === 'connections'
-            ? 'connections'
-            : 'overview'
-        : 'overview';
-    const nextProfileConnectionsView =
-      routeSection === 'profile' && requestedProfileMode === 'connections'
-        ? isProfileConnectionsView(requestedConnectionsView)
-          ? requestedConnectionsView
-          : 'following'
-        : shellChromeState.profileConnectionsView;
-
-    if (
-      shellChromeState.activePrimarySection !== routeSection ||
-      shellChromeState.timelineView !== nextTimelineView ||
-      shellChromeState.activeSettingsSection !== nextSettingsSection ||
-      shellChromeState.settingsOpen !== nextSettingsOpen ||
-      shellChromeState.profileMode !== nextProfileMode ||
-      shellChromeState.profileConnectionsView !== nextProfileConnectionsView
-    ) {
-      setShellChromeState((current) => ({
-        ...current,
-        activePrimarySection: routeSection,
-        timelineView: nextTimelineView,
-        activeSettingsSection: nextSettingsSection as SettingsSection,
-        settingsOpen: nextSettingsOpen,
-        profileMode: nextProfileMode,
-        profileConnectionsView: nextProfileConnectionsView,
-      }));
-    }
-
-    if (requestedTimelineView && requestedTimelineView !== 'bookmarks') {
-      shouldNormalize = true;
-    }
-    if (requestedTimelineView && routeSection !== 'timeline') {
-      shouldNormalize = true;
-    }
-    if (requestedSettingsSection && !isSettingsSection(requestedSettingsSection)) {
-      shouldNormalize = true;
-    }
-    if (
-      requestedProfileMode &&
-      requestedProfileMode !== 'edit' &&
-      requestedProfileMode !== 'connections'
-    ) {
-      shouldNormalize = true;
-    }
-    if (requestedProfileMode && routeSection !== 'profile') {
-      shouldNormalize = true;
-    }
-    if (
-      requestedConnectionsView &&
-      (requestedProfileMode !== 'connections' ||
-        !isProfileConnectionsView(requestedConnectionsView))
-    ) {
-      shouldNormalize = true;
-    }
-    if (routeSection === 'messages' && requestedContext) {
-      shouldNormalize = true;
-    }
-    if (
-      routeSection === 'notifications' &&
-      (requestedTimelineView ||
-        requestedChannelParam ||
-        requestedContext ||
-        requestedProfileMode ||
-        requestedConnectionsView ||
-        requestedThreadId ||
-        requestedFocusObjectId ||
-        requestedAuthorPubkey ||
-        requestedPeerPubkey ||
-        requestedSessionId ||
-        requestedRoomId)
-    ) {
-      shouldNormalize = true;
-    }
-
-    if (nextTimelineView === 'bookmarks') {
-      normalizedSelectedThread = null;
-      normalizedFocusedObjectId = null;
-      normalizedSelectedAuthorPubkey = null;
-      normalizedSelectedDirectMessagePeerPubkey = null;
-      if (requestedContext) {
-        shouldNormalize = true;
-      }
-      if (requestedFocusObjectId || requestedSessionId || requestedRoomId) {
-        shouldNormalize = true;
-      }
-      if (selectedThread) {
-        setSelectedThread(null);
-        setFocusedObjectId(null);
-        setThread([]);
-        setReplyTarget(null);
-        setRepostTarget(null);
-      }
-      if (focusedObjectId) {
-        setFocusedObjectId(null);
-      }
-      if (selectedAuthorPubkey) {
-        setSelectedAuthorPubkey(null);
-        setSelectedAuthor(null);
-        setAuthorError(null);
-      }
-      if (directMessagePaneOpen) {
-        setDirectMessagePaneOpen(false);
-      }
-      if (selectedDirectMessagePeerPubkey) {
-        setSelectedDirectMessagePeerPubkey(null);
-      }
-      setDirectMessageError(null);
-    }
-    if (routeSection === 'messages') {
-      normalizedSelectedThread = null;
-      normalizedFocusedObjectId = null;
-      if (requestedThreadId) {
-        shouldNormalize = true;
-      }
-      if (requestedFocusObjectId || requestedSessionId || requestedRoomId) {
-        shouldNormalize = true;
-      }
-      if (selectedThread) {
-        setSelectedThread(null);
-        setFocusedObjectId(null);
-        setThread([]);
-        setReplyTarget(null);
-        setRepostTarget(null);
-      }
-      if (focusedObjectId) {
-        setFocusedObjectId(null);
-      }
-      if (!directMessagePaneOpen) {
-        setDirectMessagePaneOpen(true);
-      }
-      if (!requestedPeerPubkey) {
-        normalizedSelectedDirectMessagePeerPubkey = null;
-        if (selectedDirectMessagePeerPubkey) {
-          setSelectedDirectMessagePeerPubkey(null);
-        }
-        setDirectMessageError(null);
-      } else if (!isHex64(requestedPeerPubkey)) {
-        shouldNormalize = true;
-        normalizedSelectedDirectMessagePeerPubkey = null;
-        if (selectedDirectMessagePeerPubkey) {
-          setSelectedDirectMessagePeerPubkey(null);
-        }
-      } else if (
-        requestedPeerPubkey !== selectedDirectMessagePeerPubkey ||
-        !directMessagePaneOpen
-      ) {
-        normalizedSelectedDirectMessagePeerPubkey = requestedPeerPubkey;
-        void openDirectMessagePane(requestedPeerPubkey, {
-          historyMode: 'replace',
-          normalizeOnError: true,
-          preserveAuthorPane: requestedAuthorPubkey !== null && isHex64(requestedAuthorPubkey),
-          preservedAuthorPubkey:
-            requestedAuthorPubkey && isHex64(requestedAuthorPubkey)
-              ? requestedAuthorPubkey
-              : null,
-        });
-      } else {
-        normalizedSelectedDirectMessagePeerPubkey = requestedPeerPubkey;
-      }
-      if (!requestedAuthorPubkey) {
-        normalizedSelectedAuthorPubkey = null;
-        if (selectedAuthorPubkey) {
-          setSelectedAuthorPubkey(null);
-          setSelectedAuthor(null);
-          setAuthorError(null);
-        }
-      } else if (!isHex64(requestedAuthorPubkey)) {
-        shouldNormalize = true;
-        normalizedSelectedAuthorPubkey = null;
-        if (selectedAuthorPubkey) {
-          setSelectedAuthorPubkey(null);
-          setSelectedAuthor(null);
-          setAuthorError(null);
-        }
-      } else if (
-        requestedAuthorPubkey !== selectedAuthorPubkey ||
-        !selectedAuthor ||
-        (requestedPeerPubkey ?? null) !== (selectedDirectMessagePeerPubkey ?? null)
-      ) {
-        normalizedSelectedAuthorPubkey = requestedAuthorPubkey;
-        void openAuthorDetail(requestedAuthorPubkey, {
-          historyMode: 'replace',
-          normalizeOnError: true,
-          preserveDirectMessageContext: true,
-          directMessagePeerPubkey:
-            requestedPeerPubkey && isHex64(requestedPeerPubkey) ? requestedPeerPubkey : null,
-        });
-      } else {
-        normalizedSelectedAuthorPubkey = requestedAuthorPubkey;
-      }
-    } else if (routeSection === 'notifications') {
-      normalizedSelectedThread = null;
-      normalizedFocusedObjectId = null;
-      normalizedSelectedAuthorPubkey = null;
-      normalizedSelectedDirectMessagePeerPubkey = null;
-      if (selectedThread) {
-        setSelectedThread(null);
-        setFocusedObjectId(null);
-        setThread([]);
-        setReplyTarget(null);
-        setRepostTarget(null);
-      }
-      if (focusedObjectId) {
-        setFocusedObjectId(null);
-      }
-      if (selectedAuthorPubkey) {
-        setSelectedAuthorPubkey(null);
-        setSelectedAuthor(null);
-        setAuthorError(null);
-      }
-      if (directMessagePaneOpen || selectedDirectMessagePeerPubkey) {
-        setDirectMessagePaneOpen(false);
-        setSelectedDirectMessagePeerPubkey(null);
-        setDirectMessageError(null);
-      }
-    } else if (
-      routeSection === 'timeline' &&
-      nextTimelineView !== 'bookmarks' &&
-      requestedContext === 'thread'
-    ) {
-      normalizedSelectedDirectMessagePeerPubkey = null;
-      const threadReadyForNestedAuthor =
-        requestedThreadId !== null &&
-        requestedThreadId.length > 0 &&
-        requestedThreadId === selectedThread &&
-        thread.length > 0;
-
-      if (!requestedThreadId) {
-        shouldNormalize = true;
-        normalizedSelectedThread = null;
-        normalizedFocusedObjectId = null;
-        normalizedSelectedAuthorPubkey = null;
-        if (selectedThread || selectedAuthorPubkey) {
-          setSelectedThread(null);
-          setFocusedObjectId(null);
-          setThread([]);
-          setReplyTarget(null);
-          setRepostTarget(null);
-          setSelectedAuthorPubkey(null);
-          setSelectedAuthor(null);
-          setAuthorError(null);
-        }
-      } else if (requestedThreadId !== selectedThread || thread.length === 0) {
-        normalizedSelectedThread = requestedThreadId;
-        normalizedFocusedObjectId = requestedFocusObjectId;
-        void openThread(requestedThreadId, {
-          focusObjectId: requestedFocusObjectId,
-          historyMode: 'replace',
-          normalizeOnEmpty: true,
-          topic: nextTopic,
-        });
-      } else {
-        normalizedSelectedThread = requestedThreadId;
-        if (!requestedFocusObjectId) {
-          normalizedFocusedObjectId = null;
-          if (focusedObjectId) {
-            setFocusedObjectId(null);
-          }
-        } else if (thread.some((item) => item.object_id === requestedFocusObjectId)) {
-          normalizedFocusedObjectId = requestedFocusObjectId;
-          if (focusedObjectId !== requestedFocusObjectId) {
-            setFocusedObjectId(requestedFocusObjectId);
-          }
-        } else {
-          shouldNormalize = true;
-          normalizedFocusedObjectId = null;
-          if (focusedObjectId) {
-            setFocusedObjectId(null);
-          }
-        }
-      }
-      if (!requestedAuthorPubkey) {
-        normalizedSelectedAuthorPubkey = null;
-        if (selectedAuthorPubkey) {
-          setSelectedAuthorPubkey(null);
-          setSelectedAuthor(null);
-          setAuthorError(null);
-        }
-      } else if (!isHex64(requestedAuthorPubkey)) {
-        shouldNormalize = true;
-        normalizedSelectedAuthorPubkey = null;
-        if (selectedAuthorPubkey) {
-          setSelectedAuthorPubkey(null);
-          setSelectedAuthor(null);
-          setAuthorError(null);
-        }
-      } else if (!threadReadyForNestedAuthor) {
-        normalizedSelectedAuthorPubkey = null;
-        if (selectedAuthorPubkey) {
-          setSelectedAuthorPubkey(null);
-          setSelectedAuthor(null);
-          setAuthorError(null);
-        }
-      } else if (
-        requestedAuthorPubkey !== selectedAuthorPubkey ||
-        !selectedAuthor ||
-        requestedThreadId !== selectedThread
-      ) {
-        normalizedSelectedAuthorPubkey = requestedAuthorPubkey;
-        void openAuthorDetail(requestedAuthorPubkey, {
-          fromThread: true,
-          historyMode: 'replace',
-          normalizeOnError: true,
-          threadId: requestedThreadId,
-        });
-      } else {
-        normalizedSelectedAuthorPubkey = requestedAuthorPubkey;
-      }
-    } else if (
-      routeSection === 'timeline' &&
-      nextTimelineView !== 'bookmarks' &&
-      requestedContext === 'author'
-    ) {
-      normalizedSelectedThread = null;
-      normalizedFocusedObjectId = null;
-      normalizedSelectedDirectMessagePeerPubkey = null;
-      if (requestedThreadId) {
-        shouldNormalize = true;
-      }
-      if (requestedFocusObjectId || requestedSessionId || requestedRoomId) {
-        shouldNormalize = true;
-      }
-      if (selectedThread) {
-        setSelectedThread(null);
-        setFocusedObjectId(null);
-        setThread([]);
-        setReplyTarget(null);
-        setRepostTarget(null);
-      }
-      if (focusedObjectId) {
-        setFocusedObjectId(null);
-      }
-      if (!requestedAuthorPubkey) {
-        shouldNormalize = true;
-        normalizedSelectedAuthorPubkey = null;
-        if (selectedAuthorPubkey) {
-          setSelectedAuthorPubkey(null);
-          setSelectedAuthor(null);
-          setAuthorError(null);
-        }
-      } else if (!isHex64(requestedAuthorPubkey)) {
-        shouldNormalize = true;
-        normalizedSelectedAuthorPubkey = null;
-        if (selectedAuthorPubkey) {
-          setSelectedAuthorPubkey(null);
-          setSelectedAuthor(null);
-          setAuthorError(null);
-        }
-      } else if (requestedAuthorPubkey !== selectedAuthorPubkey || !selectedAuthor) {
-        normalizedSelectedAuthorPubkey = requestedAuthorPubkey;
-        void openAuthorDetail(requestedAuthorPubkey, {
-          historyMode: 'replace',
-          normalizeOnError: true,
-        });
-      } else {
-        normalizedSelectedAuthorPubkey = requestedAuthorPubkey;
-      }
-    } else if (routeSection === 'live') {
-      normalizedSelectedThread = null;
-      normalizedFocusedObjectId = null;
-      normalizedSelectedAuthorPubkey = null;
-      normalizedSelectedDirectMessagePeerPubkey = null;
-      if (
-        requestedContext ||
-        requestedThreadId ||
-        requestedFocusObjectId ||
-        requestedAuthorPubkey ||
-        requestedPeerPubkey ||
-        requestedRoomId
-      ) {
-        shouldNormalize = true;
-      }
-      if (selectedThread) {
-        setSelectedThread(null);
-        setFocusedObjectId(null);
-        setThread([]);
-        setReplyTarget(null);
-        setRepostTarget(null);
-      }
-      if (focusedObjectId) {
-        setFocusedObjectId(null);
-      }
-      if (selectedAuthorPubkey) {
-        setSelectedAuthorPubkey(null);
-        setSelectedAuthor(null);
-        setAuthorError(null);
-      }
-      if (directMessagePaneOpen || selectedDirectMessagePeerPubkey) {
-        setDirectMessagePaneOpen(false);
-        setSelectedDirectMessagePeerPubkey(null);
-        setDirectMessageError(null);
-      }
-      normalizedSelectedGameRoomId = null;
-      if (selectedGameRoomId) {
-        setSelectedGameRoomId(null);
-      }
-      if (!requestedSessionId) {
-        normalizedSelectedLiveSessionId = null;
-        if (selectedLiveSessionId) {
-          setSelectedLiveSessionId(null);
-        }
-      } else if (
-        livePanelState?.status === 'ready' &&
-        !liveSessionsForTopic.some((session) => session.session_id === requestedSessionId)
-      ) {
-        shouldNormalize = true;
-        normalizedSelectedLiveSessionId = null;
-        if (selectedLiveSessionId) {
-          setSelectedLiveSessionId(null);
-        }
-      } else {
-        normalizedSelectedLiveSessionId = requestedSessionId;
-        if (selectedLiveSessionId !== requestedSessionId) {
-          setSelectedLiveSessionId(requestedSessionId);
-        }
-      }
-    } else if (routeSection === 'game') {
-      normalizedSelectedThread = null;
-      normalizedFocusedObjectId = null;
-      normalizedSelectedAuthorPubkey = null;
-      normalizedSelectedDirectMessagePeerPubkey = null;
-      if (
-        requestedContext ||
-        requestedThreadId ||
-        requestedFocusObjectId ||
-        requestedAuthorPubkey ||
-        requestedPeerPubkey ||
-        requestedSessionId
-      ) {
-        shouldNormalize = true;
-      }
-      if (selectedThread) {
-        setSelectedThread(null);
-        setFocusedObjectId(null);
-        setThread([]);
-        setReplyTarget(null);
-        setRepostTarget(null);
-      }
-      if (focusedObjectId) {
-        setFocusedObjectId(null);
-      }
-      if (selectedAuthorPubkey) {
-        setSelectedAuthorPubkey(null);
-        setSelectedAuthor(null);
-        setAuthorError(null);
-      }
-      if (directMessagePaneOpen || selectedDirectMessagePeerPubkey) {
-        setDirectMessagePaneOpen(false);
-        setSelectedDirectMessagePeerPubkey(null);
-        setDirectMessageError(null);
-      }
-      normalizedSelectedLiveSessionId = null;
-      if (selectedLiveSessionId) {
-        setSelectedLiveSessionId(null);
-      }
-      if (!requestedRoomId) {
-        normalizedSelectedGameRoomId = null;
-        if (selectedGameRoomId) {
-          setSelectedGameRoomId(null);
-        }
-      } else if (
-        gamePanelState?.status === 'ready' &&
-        !gameRoomsForTopic.some((room) => room.room_id === requestedRoomId)
-      ) {
-        shouldNormalize = true;
-        normalizedSelectedGameRoomId = null;
-        if (selectedGameRoomId) {
-          setSelectedGameRoomId(null);
-        }
-      } else {
-        normalizedSelectedGameRoomId = requestedRoomId;
-        if (selectedGameRoomId !== requestedRoomId) {
-          setSelectedGameRoomId(requestedRoomId);
-        }
-      }
-    } else if (routeSection === 'timeline' && nextTimelineView !== 'bookmarks' && requestedContext) {
-      shouldNormalize = true;
-      normalizedSelectedThread = null;
-      normalizedFocusedObjectId = null;
-      normalizedSelectedAuthorPubkey = null;
-      normalizedSelectedDirectMessagePeerPubkey = null;
-      if (selectedThread || selectedAuthorPubkey) {
-        setSelectedThread(null);
-        setFocusedObjectId(null);
-        setThread([]);
-        setReplyTarget(null);
-        setRepostTarget(null);
-        setSelectedAuthorPubkey(null);
-        setSelectedAuthor(null);
-        setAuthorError(null);
-      }
-      if (focusedObjectId) {
-        setFocusedObjectId(null);
-      }
-      if (directMessagePaneOpen || selectedDirectMessagePeerPubkey) {
-        setDirectMessagePaneOpen(false);
-        setSelectedDirectMessagePeerPubkey(null);
-        setDirectMessageError(null);
-      }
-    } else {
-      if (
-        requestedThreadId ||
-        requestedFocusObjectId ||
-        requestedAuthorPubkey ||
-        requestedPeerPubkey ||
-        requestedSessionId ||
-        requestedRoomId
-      ) {
-        shouldNormalize = true;
-      }
-      normalizedSelectedThread = null;
-      normalizedFocusedObjectId = null;
-      normalizedSelectedAuthorPubkey = null;
-      normalizedSelectedDirectMessagePeerPubkey = null;
-      if (
-        selectedThread ||
-        focusedObjectId ||
-        selectedAuthorPubkey ||
-        directMessagePaneOpen ||
-        selectedDirectMessagePeerPubkey
-      ) {
-        setSelectedThread(null);
-        setFocusedObjectId(null);
-        setThread([]);
-        setReplyTarget(null);
-        setRepostTarget(null);
-        setSelectedAuthorPubkey(null);
-        setSelectedAuthor(null);
-        setAuthorError(null);
-        setDirectMessagePaneOpen(false);
-        setSelectedDirectMessagePeerPubkey(null);
-        setDirectMessageError(null);
-      }
-    }
-
-    if (shouldReload) {
-      void loadTopics(
-        trackedTopics,
-        nextTopic,
-        requestedContext === 'thread' ? requestedThreadId : null
-      ).catch(() => undefined);
-    }
-
-    if (shouldNormalize) {
-      scheduleAnimationFrame(() => {
-        syncRoute('replace', {
-          activeTopic: nextTopic,
-          composeTarget: privateComposeTarget(nextSelectedChannelId),
-          focusedObjectId: normalizedFocusedObjectId,
-          primarySection: routeSection,
-          profileConnectionsView: nextProfileConnectionsView,
-          profileMode: nextProfileMode,
-          selectedAuthorPubkey: normalizedSelectedAuthorPubkey,
-          selectedDirectMessagePeerPubkey: normalizedSelectedDirectMessagePeerPubkey,
-          selectedGameRoomId: normalizedSelectedGameRoomId,
-          selectedLiveSessionId: normalizedSelectedLiveSessionId,
-          selectedThread: normalizedSelectedThread,
-          settingsOpen: nextSettingsOpen,
-          settingsSection: nextSettingsSection as SettingsSection,
-          timelineScope: privateTimelineScope(nextSelectedChannelId),
-          timelineView: nextTimelineView,
-        });
-      });
-    }
-  }, [
-    activeTopic,
-    directMessagePaneOpen,
-    focusedObjectId,
-    gamePanelStateByTopic,
-    gameRoomsByTopic,
-    joinedChannelsByTopic,
+  useRouteSynchronization({
     loadTopics,
-    livePanelStateByTopic,
-    liveSessionsByTopic,
+    lastObservedRouteUrlRef,
     navigate,
     openAuthorDetail,
     openDirectMessagePane,
     openThread,
     pendingRouteUrlRef,
+    resolvedRouteLocation,
     routeSection,
-    resolvedRouteLocation.pathname,
-    resolvedRouteLocation.search,
     scheduleAnimationFrame,
-    selectedAuthor,
-    selectedAuthorPubkey,
-    selectedChannelIdByTopic,
-    selectedDirectMessagePeerPubkey,
-    selectedGameRoomId,
-    selectedLiveSessionId,
-    selectedThread,
-    setActiveTopic,
-    setAuthorError,
-    setComposeChannelByTopic,
-    setDirectMessageError,
-    setDirectMessagePaneOpen,
-    setFocusedObjectId,
-    setReplyTarget,
-    setRepostTarget,
-    setSelectedAuthor,
-    setSelectedAuthorPubkey,
-    setSelectedChannelIdByTopic,
-    setSelectedDirectMessagePeerPubkey,
-    setSelectedGameRoomId,
-    setSelectedLiveSessionId,
-    setSelectedThread,
-    setShellChromeState,
-    setThread,
-    setTimelineScopeByTopic,
-    setLastNonNotificationsRoute,
-    shellChromeState.activePrimarySection,
-    shellChromeState.activeSettingsSection,
-    shellChromeState.profileConnectionsView,
-    shellChromeState.profileMode,
-    shellChromeState.settingsOpen,
-    shellChromeState.timelineView,
+    state,
     syncRoute,
-    thread,
-    trackedTopics,
-  ]);
+    storeApi,
+  });
 
   return {
     routeSection,
